# docs/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Payment Operations Investigator — Public Java API",
    "version": "0.1.0",
    "description": "Source-derived local synthetic portfolio API. Session-bound tenant/role authorization, immutable investigation evidence and human review; no real-payment commands. Only public Java routes are documented, not internal Python worker endpoints. Request DTOs reject unknown fields. Response metadata that Java deliberately leaves extensible remains optional. Monetary values use exact JSON-safe integer INR minor units. HTTP 200 is used for successful login, investigation and decision creation. This document does not claim production authentication, live-model success or database-administrator tamper resistance. Implements bounded original-synthetic OBPM NEFT ECA evidence imports; no Oracle source connector or banking-side commands. Import-specific validation and error status differ from ordinary DTO routes. A separate local mock inquiry transport now fetches original synthetic ECA evidence; it is not an Oracle product connector."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8088",
      "description": "Local Java API; web deployments normally proxy /api on the same origin."
    }
  ],
  "security": [
    {
      "SessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "operationId": "dashboard",
        "tags": [
          "System"
        ],
        "summary": "Read tenant totals and recent activity",
        "description": "Counts and total cover all visible cases. Java uses checked integer arithmetic and rejects an unsafe aggregate rather than returning a rounded or wrapped amount. These are imported-case metrics, not all-payment OBPM analytics or success rates.",
        "x-roles": [
          "ANALYST",
          "REVIEWER",
          "VIEWER"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "422": {
            "$ref": "#/components/responses/InvalidData"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "POI_SESSION",
        "description": "Opaque HttpOnly SameSite=Strict cookie obtained via POST /api/auth/login. Authenticated mutations also require the separate X-CSRF-Token header. Do not send tenant/role headers as authorization."
      }
    },
...
    "responses": {
...
      "Unauthenticated": {
        "description": "401: no valid authenticated session, or invalid login credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
      "InvalidData": {
        "description": "422: unsupported/invalid stored monetary data, mixed currencies, totals outside the JSON-safe range, or invalid stored proposal. No partial monetary result is returned.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "500: unexpected server/framework error; response does not expose stack traces.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "CaseAmount": {
        "type": "integer",
        "minimum": 0,
        "maximum": 9007199254740991,
        "description": "Nonnegative original case amount, in INR minor units."
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable application/security error code. The examples are not an exhaustive enum.",
            "examples": [
              "INVALID_REQUEST",
              "UNAUTHENTICATED",
              "INVALID_CREDENTIALS",
              "FORBIDDEN",
              "NOT_FOUND",
              "MAKER_CHECKER_REQUIRED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_KEY_REQUIRED",
              "IDEMPOTENCY_CONFLICT",
              "ALREADY_REVIEWED",
              "AMOUNT_OUT_OF_RANGE",
              "CURRENCY_MISMATCH",
              "WORKER_UNAVAILABLE",
              "INVALID_WORKER_RESPONSE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "Safe client-facing message; no database or provider stack trace."
          },
          "requestId": {
            "type": "string",
            "description": "Server-generated request correlation identifier."
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "RecentActivity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "description": "Current values: FIXTURE_IMPORTED, INVESTIGATION_CREATED, INVESTIGATION_FAILED, REVIEW_APPROVE, REVIEW_REJECT."
          },
          "detail": {
            "type": "string"
          },
          "caseId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "occurredAt",
          "actor",
          "action",
          "detail",
          "caseId"
        ],
        "additionalProperties": false
      },
...
      "Dashboard": {
        "type": "object",
        "properties": {
          "openCases": {
            "type": "integer",
            "minimum": 0,
            "description": "All tenant cases whose status is not RESOLVED, including awaiting review/escalated/needs evidence."
          },
          "highPriorityCases": {
            "type": "integer",
            "minimum": 0,
            "description": "HIGH-priority cases whose status is not RESOLVED."
          },
          "awaitingReview": {
            "type": "integer",
            "minimum": 0
          },
          "resolvedCases": {
            "type": "integer",
            "minimum": 0
          },
          "totalAmountMinor": {
            "$ref": "#/components/schemas/CaseAmount",
            "description": "Sum of all tenant case amounts, including resolved cases. Out-of-range aggregate returns 422."
          },
          "currency": {
            "type": "string",
            "const": "INR"
          },
          "recentActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentActivity"
            },
            "maxItems": 10
          },
          "mode": {
            "type": "string",
            "const": "synthetic"
          }
        },
        "required": [
          "openCases",
          "highPriorityCases",
          "awaitingReview",
          "resolvedCases",
          "totalAmountMinor",
          "currency",
          "recentActivity",
          "mode"
        ],
        "additionalProperties": false,
        "description": "Imported-case workload analytics only. Counts and summed case amounts include synthetic fixture cases and imported OBPM cases; they do not measure all OBPM payments, payment success rate or settlement throughput."
      },
...
    }
  },
...
}